Refactoring: Following Restful Project Structure from its doc
