+++ api.py
@@ -1,13 +1,15 @@
 from flask import Flask, jsonify, request
-from flask_restful import reqparse, abort, Api, Resource
+from flask_restful import reqparse, abort, Api
 from werkzeug.security import safe_str_cmp
 from flask_jwt_extended import (
     JWTManager, jwt_required, create_access_token,
     get_jwt_identity
 )
 
-from model.User import User
-from model.Link import Link
+from common.Link import Link
+from common.User import User
+
+from resources.TodoList import TodoList
 
 app = Flask(__name__)
 api = Api(app)
@@ -48,7 +50,7 @@ def login():
 
     user = username_table.get(username, None)
     if user and safe_str_cmp(user.password.encode('utf-8'), password.encode('utf-8')):
-        # Identity can be any data that is json serializable
+        #Identity can be any data that is json serializable
         access_token = create_access_token(identity=username)
         return jsonify(access_token=access_token), 200
     else:
@@ -64,61 +66,10 @@ def protected():
     return jsonify(logged_in_as=current_user), 200
 
 
-TODOS = {
-    'todo1': {'task': 'build an API'},
-    'todo2': {'task': '?????'},
-    'todo3': {'task': 'profit!'},
-}
-
-
-def abort_if_todo_doesnt_exist(todo_id):
-    if todo_id not in TODOS:
-        abort(404, message="Todo {} doesn't exist".format(todo_id))
-
-parser = reqparse.RequestParser()
-parser.add_argument('task')
-
-
-# Todo
-# shows a single todo item and lets you delete a todo item
-class Todo(Resource):
-    @jwt_required
-    def get(self, todo_id):
-        abort_if_todo_doesnt_exist(todo_id)
-        return TODOS[todo_id]
-
-    @jwt_required
-    def delete(self, todo_id):
-        abort_if_todo_doesnt_exist(todo_id)
-        del TODOS[todo_id]
-        return '', 204
-
-    @jwt_required
-    def put(self, todo_id):
-        args = parser.parse_args()
-        task = {'task': args['task']}
-        TODOS[todo_id] = task
-        return task, 201
-# TodoList
-# shows a list of all todos, and lets you POST to add new tasks
-class TodoList(Resource):
-    @jwt_required    
-    def get(self):
-        return TODOS
-
-    @jwt_required
-    def post(self):
-        args = parser.parse_args()
-        todo_id = int(max(TODOS.keys()).lstrip('todo')) + 1
-        todo_id = 'todo%i' % todo_id
-        TODOS[todo_id] = {'task': args['task']}
-        return TODOS[todo_id], 201
-
 ##
 ## Actually setup the Api resource routing here
 ##
 api.add_resource(TodoList, '/todos')
-api.add_resource(Todo, '/todos/<todo_id>')
 
 
 if __name__ == '__main__':
